type cast the volume and accumulated volume to integers to avoid the default output format

+++ src/Ada/Root_Counts/DEMiCs/simplex.cpp
@@ -3796,8 +3796,8 @@ void simplex::calMixedVol ( lvData* lv, int* sp, int supN )
    mixedVol += det;
 
 #ifdef compile4phc
-   strcell << "volume : " << det;
-   strcell << " accumulated volume : " << mixedVol;
+   strcell << "volume : " << ((int) det);
+   strcell << " accumulated volume : " << ((int) mixedVol);
    // cout << "strcell: " << strcell.str() << endl;
    int fail = demics_append_cell_indices(strcell.str());
 #endif
